requirements.txt
Fabric==1.10.2
argparse==1.2.1
ecdsa==0.13
paramiko==1.15.2
pycrypto==2.6.1
wsgiref==0.1.2
yolk==0.4.3
PyYAML==3.11
pika==0.9.14
